Show current buffer first, if not shown already
When a buffer is already show in the tabline, it should keep its
position. But if a valid buffer was not shown because not recent enough,
show it first when it is opened again. The same happens for newly opened
valid buffers, they are moved on top of the ordered list.